requirements.txt
# Core dependencies
manim>=0.18.0
numpy>=1.21.0
matplotlib>=3.5.0

# AI/ML libraries
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
requests>=2.28.0

# Audio processing
pydub>=0.25.1
librosa>=0.10.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.6.0

# Data handling
pandas>=1.5.0
json5>=0.9.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0
click>=8.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
pytest>=7.0.0

# Optional: For advanced features
# ffmpeg-python>=0.2.0
# scipy>=1.9.0
# scikit-learn>=1.1.0 